Applied suggested thematic grouping in api handlers and cli, removed some line breaks
